actualizar

+++ app/Http/Livewire/GestionarClientes.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Livewire;
 
+use App\Clases\UserMoodle;
+use App\Models\Cmoodle;
 use App\Models\User;
 use Illuminate\Support\Facades\DB;
 use Livewire\Component;
@@ -10,9 +12,10 @@ class GestionarClientes extends Component
 {
     public User $cliente;
     public $modal_titulo;
+    public $sdominio;
     public $bcliente;
     public $gclientes;
-
+    protected $listeners = ['modal'];
     protected $rules = [
         'cliente.name' => 'required|string',
         'cliente.ap_paterno' => '',
@@ -30,6 +33,11 @@ class GestionarClientes extends Component
         'cliente.celular' => 'Celular',
     ];
 
+    public function mount(){
+        $cmoodle = Cmoodle::find(1);
+        $this->sdominio = $cmoodle->url;
+    }
+
     #buscar listado de estudiante
     public function updatedBcliente()
     {
@@ -39,7 +47,7 @@ class GestionarClientes extends Component
                                 ->orwhere('name', 'like', '%' . $this->bcliente.'%')
                                 ->orwhere('ap_paterno', 'like', '%' . $this->bcliente.'%')
                                 ->orwhere('ap_materno', 'like', '%' . $this->bcliente.'%');
-                    })->orderBy('id', 'DESC')->get();
+                    })->WhereNull('deleted_at')->orderBy('id', 'DESC')->get();
     }
     #obtener datos
     public function obtener_datos($usuario){
@@ -58,30 +66,45 @@ class GestionarClientes extends Component
     }
     #eliminar
     public function eliminar(User $cliente){
-        if($cliente->comprobantes->count() == 0){
+        if($cliente->comprobantes->count() == 0 && $cliente->cmatriculas->count() == 0){
+            $moodle = UserMoodle::buscar($cliente->id);
+            $moodle->eliminar();
             $cliente->delete();
             $this->updatedBcliente();
         }
+        else{$this->emit('notificar_error_usuario','el usuario no se puede eliminar');}
     }
-
+    #saved
     public function save($accion){
         if($accion == "Crear"){
             $this->validate();
-
         $this->cliente->password = bcrypt($this->cliente->email);
         }
         else {
             $this->validate(['cliente.email'=>'required|unique:users,email,'.$this->cliente->id]);
         }
         $this->cliente->save();
+
+        #crear el usuario en moodle
+        if($accion == "Crear")
+        {
+            $moodle = new UserMoodle($this->cliente->id);
+            $n_user = $moodle->crear_usuario();
+            if ($n_user) {$this->emit('notificar_listado');$this->updatedBcliente();}
+            else{$this->emit('notificar_error_usuario','no se creo el usuario, verificar los datos escritos');}
+        }
+
+        else{
+            $moodle = UserMoodle::buscar($this->cliente->id);
+            $moodle->modificar_usuario();
+            $this->emit('notificar_listado');$this->updatedBcliente();
+        }
+
         $this->modal_titulo = "Crear";
         $this->cliente = new User();
-        $this->emit('notificar_listado');
-        $this->updatedBcliente();
+
     }
-
-    protected $listeners = ['modal'];
-
+    #render
     public function render()
     {
         return view('livewire.gestionar-clientes');
